Cleanup, inline load-photoswipe.js to allow JS config via SiteParam variables
Old, useless files removed:
- footer.html (my own experiment; Hugo DOES NOT combine multiple footers, it picks one or the other)
- load-photoswipe-theme.html (duplicate from the original repo)

load-photoswipe has been inlined so that it accepts {{ if .Site.Params.debug }} variable to silence the console logging of image dimensions. There are a couple of other ways to pass variables to JS, where your JS remains a separate file and is not inlined:
- js.Build method (https://discourse.gohugo.io/t/variables-inside-javascript/39086/4)
- ExecuteAsTemplate method (https://discourse.gohugo.io/t/inserting-templates-variables-in-javascript/4246/6)

But I couldn't get either to work.