Close /gallery tag (fix Li-Wen Yip's example)
The original author provided an incorrect example (https://www.liwen.id.au/heg/#Create-gallery-from-a-directory) where he didn't close the {{< /gallery >}} tag. If this is not done, the webpage breaks. I was sitting on this error for a while and just managed to fix it while tinkering with (what I thought was) a CSS load error.